refactor: move multi runner test to runner
